+++ pallets/loans/benchmarking/src/mock.rs
@@ -257,9 +257,9 @@ pub fn new_test_ext() -> sp_io::TestExternalities {
         borrow_index: Rate::one(),                             // 1
         exchange_rate: Rate::saturating_from_rational(2, 100), // 0.02
         base_rate: Rate::saturating_from_rational(2, 100),     // 2%
-        kink_rate: Rate::saturating_from_rational(10, 100),    // 10%
+        jump_rate: Rate::saturating_from_rational(10, 100),    // 10%
         full_rate: Rate::saturating_from_rational(32, 100),    // 32%
-        kink_utilization: Ratio::from_percent(80),             // 80%
+        jump_utilization: Ratio::from_percent(80),             // 80%
         collateral_factor: vec![
             (CurrencyId::DOT, Ratio::from_percent(50)),
             (CurrencyId::KSM, Ratio::from_percent(50)),